MSVC compat

+++ cpp/src/arrow/compute/kernels/scalar_cast_nested.cc
@@ -163,7 +163,7 @@ struct CastStruct {
                                batch[0].type()->ToString(), " ", out->type()->ToString());
     }
 
-    for (int64_t i = 0; i < in_field_count; ++i) {
+    for (int i = 0; i < in_field_count; ++i) {
       const auto in_field_name =
           checked_cast<const StructType&>(*batch[0].type()).field(i)->name();
       const auto out_field_name =
@@ -181,7 +181,7 @@ struct CastStruct {
 
       DCHECK(!out_scalar->is_valid);
       if (in_scalar.is_valid) {
-        for (int64_t i = 0; i < in_field_count; i++) {
+        for (int i = 0; i < in_field_count; i++) {
           auto values = in_scalar.value[i];
           auto target_type = out->type()->field(i)->type();
           ARROW_ASSIGN_OR_RAISE(Datum cast_values,
@@ -197,7 +197,7 @@ struct CastStruct {
     const ArrayData& in_array = *batch[0].array();
     ArrayData* out_array = out->mutable_array();
 
-    for (int64_t i = 0; i < in_field_count; ++i) {
+    for (int i = 0; i < in_field_count; ++i) {
       auto values = in_array.child_data[i];
       if (in_array.offset != 0) {
         values = values->Slice(in_array.offset, in_array.length);
